requirements.txt

dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
plotly==5.22.0
pandas==2.2.2
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
python-dotenv==1.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
tenacity==8.3.0
Flask==3.0.3
json5==0.9.25
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
gunicorn
requests